rust: Move Svix, SvixOptions into a separate module
… to prepare for generating api/mod.rs.
